First Thoughts
Well, this was not fun. The installation was surprisingly very simple.
But then began the trouble.
I installed the rust-analyzer extension for VSCode and, it kept giving
me "rust-analyzer failed to load workspace" error. It still does.

So, I ignored that and did the basic steps: cargo new Learning1 ->
cd Learning1 -> cargo build -> cargo run. It worked in the terminal.
I was happy.

Then, I made changes to the main.rs file, did the cargo build and it
FAILED! :(

That is where I am right now, let's see what this holds in store for me.